Add loading, splash, and menu screen

+++ desktop/src/com/gofficer/codenames/desktop/DesktopLauncher.java
@@ -8,8 +8,8 @@ public class DesktopLauncher {
 	public static void main (String[] arg) {
 		LwjglApplicationConfiguration config = new LwjglApplicationConfiguration();
 		config.title = Application.TITLE + " v" + Application.VERSION;
-		config.width = Application.V_WIDTH;
-		config.height = Application.V_HEIGHT;
+		config.width = (int) Application.V_WIDTH;
+		config.height = (int) Application.V_HEIGHT;
 		config.backgroundFPS = 60;
 		config.foregroundFPS = 60;
 		config.resizable = false;
